Adding rubocop and fixing offenses

+++ db/migrate/20150723220742_deduplicate_calibration_tags_and_user_ids_integers.rb
@@ -1,11 +1,11 @@
+# frozen_string_literal: true
+
 class DeduplicateCalibrationTagsAndUserIdsIntegers < ActiveRecord::Migration[5.2]
   def up
     change_column :tags, :user_id, :integer
 
-    Tag.where('name LIKE (?)','calibration:%').each do |tag|
-      if Tag.where(user_id: tag.user_id, name:tag.name, spectrum_id:tag.spectrum_id).length > 1
-        tag.destroy
-      end
+    Tag.where('name LIKE (?)', 'calibration:%').each do |tag|
+      tag.destroy if Tag.where(user_id: tag.user_id, name: tag.name, spectrum_id: tag.spectrum_id).length > 1
     end
   end
